requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-datastore==2.16.1
requests==2.31.0
sqlite3-wrapper==0.1.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7